feat: update токена на 0 inst/ update в amoService
